add some sql md5 checks and permissions

+++ dragonfab/database.py
@@ -1,6 +1,9 @@
 import os
 
 from fabric.api import local, env, task, sudo, lcd, put, run, cd, settings, get, require
+from fabric.contrib.files import exists
+
+from dragonfab import local_md5, remote_md5
 
 def _connection_string(env, dba=False):
     conn_str = ''
@@ -17,35 +20,43 @@ def _connection_string(env, dba=False):
     conn_str += ' '
     return conn_str
 
+rdump_path = '/var/backups/dumps/latest.sql'
+
 @task
 def dump():
     """Copy of current database from the production server to dumps/latest.sql"""
     require('local_path')
+    sudo('mkdir -p %s' % os.path.dirname(rdump_path))
+
     with lcd(env.local_dir):
         local('mkdir -p dumps')
         if os.path.exists('dumps/latest.sql'):
             local('mv dumps/latest.sql dumps/latest.sql.last')
-        sudo('mkdir -p /var/backups/dumps/')
 
-        sudo('pg_dump %s > /var/backups/dumps/latest.sql' % _connection_string(env))
+        sudo('pg_dump %s > %s' % (_connection_string(env), rdump_path))
 
-        get('/var/backups/dumps/latest.sql', 'dumps/latest.sql')
-        sudo('rm /var/backups/dumps/latest.sql')
+        get(rdump_path, 'dumps/latest.sql')
+        sudo('chmod go-rwx %s' % rdump_path)
+        local('chmod o-rwx %s' % 'dumps/latest.sql')
 
 @task
 def push():
     "Recreate database from dumps/latest.sql."
     require('db_user', 'dba')
-
-    sudo('mkdir -p /var/backups/dumps/')
-    put('dumps/latest.sql', '/var/backups/dumps/latest.sql', use_sudo=True)
-    connection_string = _connection_string(env, dba=True)
-    with settings(warn_only=True):
-        run('dropdb %s' % connection_string)
-    run('createdb -O %s %s' % (env.db_user, connection_string))
-    #  When this bug is fixed: http://trac.osgeo.org/postgis/ticket/2223
-    #  we can add "-v ON_ERROR_STOP=1" to this line
-    run('psql %s -f /var/backups/dumps/latest.sql' % connection_string)
+    sudo('mkdir -p %s' % os.path.dirname(rdump_path))
+    if not exists(rdump_path) or (remote_md5(rdump_path) != local_md5('dumps/latest.sql')):
+        put('dumps/latest.sql', rdump_path, use_sudo=True)
+        connection_string = _connection_string(env, dba=True)
+        with settings(warn_only=True):
+            run('dropdb %s' % connection_string)
+        run('createdb -O %s %s' % (env.db_user, connection_string))
+        #  When this bug is fixed: http://trac.osgeo.org/postgis/ticket/2223
+        #  we can add "-v ON_ERROR_STOP=1" to this line
+        run('psql %s -f %s' % (connection_string, rdump_path))
+        sudo('chmod go-rwx %s' % rdump_path)
+        local('chmod o-rwx %s' % 'dumps/latest.sql')
+    else:
+        print "-----> remote dumpfile is the same as local - not pushing"
 
 @task
 def migrate():
